7th lab. work full

+++ Computer_Networks_Lab_7/Computer_Networks_Lab_6/Id.hpp
@@ -5,7 +5,7 @@ namespace Id
 	struct Octet
 	{
 		char* octet;
-		int value_of_octet;
+		short value_of_octet;
 		Octet(char* arr);
 		Octet();
 	};
